requirements.txt
django  (5.0.3)
psycopg2-binary  (2.9.9)
asgiref<4,>=3.7.0  (3.8.1)
sqlparse>=0.3.1  (0.4.4)  
tzdata  (2024.1)
djangorestframework

